refactor: auto-split test_ai_workflow_engine.py
