+++ tests/unit/shared_python/ai_workflow_engine/test_workflow_step.py
@@ -1,0 +1,112 @@
+"""Unit tests for shared/python/ai/workflow_engine.py."""
+
+from __future__ import annotations
+
+from unittest.mock import MagicMock
+
+import pytest
+from src.shared.python.ai.exceptions import WorkflowError
+from src.shared.python.ai.tool_registry import ToolRegistry
+from src.shared.python.ai.types import ConversationContext, ExpertiseLevel, ToolResult
+from src.shared.python.ai.workflow_engine import (
+    RecoveryStrategy,
+    StepResult,
+    StepStatus,
+    ValidationResult,
+    Workflow,
+    WorkflowEngine,
+    WorkflowExecution,
+    WorkflowStep,
+)
+
+# ---------------------------------------------------------------------------
+# Fixtures
+# ---------------------------------------------------------------------------
+
+
+@pytest.fixture
+def registry() -> ToolRegistry:
+    return ToolRegistry()
+
+
+@pytest.fixture
+def engine(registry: ToolRegistry) -> WorkflowEngine:
+    return WorkflowEngine(registry)
+
+
+@pytest.fixture
+def context() -> ConversationContext:
+    return ConversationContext(session_id="test-session")
+
+
+def _make_workflow(
+    wf_id: str = "wf1",
+    num_steps: int = 2,
+) -> Workflow:
+    wf = Workflow(id=wf_id, name="Test Workflow", description="A test workflow")
+    for i in range(num_steps):
+        wf.add_step(
+            WorkflowStep(
+                id=f"step_{i}",
+                name=f"Step {i}",
+                description=f"Step {i} description",
+            )
+        )
+    return wf
+
+
+def _make_tool_result(success: bool = True, result: object = None) -> ToolResult:
+    tr = MagicMock(spec=ToolResult)
+    tr.success = success
+    tr.result = result or {}
+    tr.error = None if success else "tool error"
+    return tr
+
+
+# ---------------------------------------------------------------------------
+# Dataclass tests
+# ---------------------------------------------------------------------------
+
+
+class TestWorkflowStep:
+    def test_ai_workflow_engine_defaults(self) -> None:
+        s = WorkflowStep(id="s1", name="Step 1", description="Desc")
+        assert s.tool_name is None
+        assert s.on_failure == RecoveryStrategy.ASK_USER
+        assert s.timeout == 300.0
+
+    def test_custom_on_failure(self) -> None:
+        s = WorkflowStep(
+            id="s1", name="S", description="D", on_failure=RecoveryStrategy.ABORT
+        )
+        assert s.on_failure == RecoveryStrategy.ABORT
+
+
+# ---------------------------------------------------------------------------
+# WorkflowEngine — registration
+# ---------------------------------------------------------------------------
+
+
+# ---------------------------------------------------------------------------
+# WorkflowEngine — start_workflow
+# ---------------------------------------------------------------------------
+
+
+# ---------------------------------------------------------------------------
+# WorkflowEngine — is_complete / get_current_step
+# ---------------------------------------------------------------------------
+
+
+# ---------------------------------------------------------------------------
+# WorkflowEngine — execute_next_step (no-tool steps)
+# ---------------------------------------------------------------------------
+
+
+# ---------------------------------------------------------------------------
+# WorkflowEngine — get_progress
+# ---------------------------------------------------------------------------
+
+
+# ---------------------------------------------------------------------------
+# WorkflowEngine — get_step_educational_content
+# ---------------------------------------------------------------------------
